requirements.txt
fastapi
uvicorn
pydantic[email]
pydantic-settings
python-dotenv
pyjwt
python-multipart
email-validator
